Complex Parser in progress checkin
Removed the multiple query file, it wasn't used
Changed the name of MockDoubleQuery to DoubleQuery
Added equals functions to classes

Created a new set of tokenizers and parsers with accompanying tests, the
complex set. These are equipped to handle parenthesized input like (a or
(b and c))

QueryParser is now an abstract class to try to prevent mixing and
matching of parser tokenizer levels - ex: complex can only be used with
complex

Made tests more streamlined with fewer constants
Used inheritance for the ComplexTokenParserTest, it reuses all the test
cases of the SimpleTokenParserTest and adds more. - DRY principle

The complex parser is still in progress, but it has achieved full
compatibility with strings intended for the simple parser.