Removed need for Query start/end characters
Removed the simple parsers and tests because they weren't serving any
function, additionally the old code was slower, harder to understand,
and more fragile.

The tokenizer and token parser can now handle queries without []
identifying the queries.

Stopped SimpleTokenParser from implementing TokenParser so it can't be
used with a QueryParser